# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Koraspond Microservices API",
        "description": "API documentation for the application endpoints.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/products-list": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Get a list of products",
                "operationId": "fcaa94fe64619a5409b36e7694ec09b1",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "description": "Request body for getting product list",
                                "properties": {},
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product list retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "draw": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "recordsTotal": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "recordsFiltered": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "pro"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "abc"
                                                    },
                                                    "price": {
                                                        "type": "string",
                                                        "example": "10.00"
                                                    },
                                                    "quantity": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-08-29T11:00:10.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-08-29T11:10:50.000000Z"
                                                    },
                                                    "image_url": {
                                                        "type": "string",
                                                        "format": "url",
                                                        "example": "http://127.0.0.1:8000/product"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "disableOrdering": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "queries": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "query": {
                                                        "type": "string",
                                                        "example": "select count(*) as aggregate from `products`"
                                                    },
                                                    "bindings": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "time": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1.93
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "input": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/products": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Save a new product",
                "operationId": "5b81ab655941816f89d3544495453a0e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "price",
                                    "quantity"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the product",
                                        "type": "string",
                                        "example": "pro"
                                    },
                                    "description": {
                                        "description": "Description of the product",
                                        "type": "string",
                                        "example": "abc"
                                    },
                                    "price": {
                                        "description": "Price of the product",
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    },
                                    "quantity": {
                                        "description": "Quantity of the product",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}